Module Biaya Tetap: - Finish EDIT and DELETE

+++ database/migrations/2016_04_25_103938_create_table_biaya_perijinan.php
@@ -19,8 +19,8 @@ class CreateTableBiayaPerijinan extends Migration
             $table->integer('jenis_biaya_perijinan')->nullable();
             $table->string('jenis_biaya_perijinan_lain', 100)->nullable();
             $table->integer('frek_satuan')->nullable();
-            $table->float('harga_satuan')->nullable();
-            $table->float('total_biaya')->nullable();
+            $table->double('harga_satuan')->nullable();
+            $table->double('total_biaya')->nullable();
 
             $table->timestamps();
             $table->softDeletes();
